% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tftFinder.R
\name{tftFinder}
\alias{tftFinder}
\title{provide an app that develops gene set representations from TF target selections}
\usage{
tftFinder()
}
\description{
provide an app that develops gene set representations from TF target selections
}
