% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trpvar1}
\alias{trpvar1}
\title{initial version of report on variants in Transient Receptor Potential Vanilloid (TRPV) gene family on chromosome 17 for NA06985}
\format{data.frame}
\usage{
trpvar1
}
\description{
initial version of report on variants in Transient Receptor Potential Vanilloid (TRPV) gene family on chromosome 17 for NA06985
}
\note{
The variant names from VCF were not preserved in the
first draft of workflow R code.  A column called 'X' is present.
}
\keyword{datasets}
